Add --download-url and use it in Makefile
--download-url lets us directly download data from a specific URL,
instead of cloning a git repo.

The NYT repo at https://github.com/nytimes/covid-19-data has gotten
pretty big and was over 10GB when I last checked. This lets us download
only the data we rely on, at the expense of re-downloading it every
time. This tradoff is worth it to me, but the old behavior is still
around.